🏷 Types 디렉토리의 issue와 reaction type을 해당api 파일로 이동함 - 응집도를 높이기 위해 데이터 응답 관련 타입들을 api 요청 함수와 같은 파일 내에서 선언했다.
