+++ src/api/reaction.ts
@@ -1,5 +1,19 @@
-import { ReactionTypes } from '@/types/reaction';
 import axios, { AxiosError } from 'axios';
+
+export type ReactionNameType =
+  | 'THUMBS_UP'
+  | 'THUMBS_DOWN'
+  | 'LAUGH'
+  | 'PARTY_POPPER'
+  | 'CONFUSED'
+  | 'HEART'
+  | 'ROCKET'
+  | 'EYES';
+
+export interface ReactionTypes {
+  name: ReactionNameType;
+  unicode: string;
+}
 
 export const getReactionData = async (): Promise<ReactionTypes[]> => {
   try {
