+++ src/components/Organisms/LabelTable/LabelItem/index.tsx
@@ -3,7 +3,7 @@ import Label from '@/components/Atoms/Label';
 import * as S from '@/components/Organisms/LabelTable/LabelItem/index.styled';
 import { COLORS } from '@/styles/theme';
 import { LabelState } from '@/stores/label';
-import { LabelTypes } from '@/types/issue';
+import { LabelTypes } from '@/api/issue/types';
 import { useNavigate } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
 import { ModalState } from '@/components/Modal';
