+++ src/hooks/useFetchReaction.ts
@@ -1,6 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { ReactionTypes } from '@/types/reaction';
-import { getReactionData } from '@/api/reaction';
+import { ReactionTypes, getReactionData } from '@/api/reaction';
 
 const useFetchReaction = () => {
   const { data: reactions } = useQuery<ReactionTypes[]>(['reactions'], getReactionData);
