--- src/types/reaction.ts
@@ -1,14 +1,0 @@
-export type ReactionNameType =
-  | 'THUMBS_UP'
-  | 'THUMBS_DOWN'
-  | 'LAUGH'
-  | 'PARTY_POPPER'
-  | 'CONFUSED'
-  | 'HEART'
-  | 'ROCKET'
-  | 'EYES';
-
-export interface ReactionTypes {
-  name: ReactionNameType;
-  unicode: string;
-}
